(IMAGES-976) Improve build parameter injection
Continue the work from IMAGES-1018 to remove environment variables to
inject data into the build:

1. Switch to .json format as this is easier to consume in powershell.
   This is done with PS2 support for the Windows-2008/2008r2 platform.
2. This requires changes to the puppet lookup and the variables used
   to access the .json data.
3. Windows Version/Edition information is passed in via .json file to
   be consumed by the packer testing.
4. This required moving the platform-package file processing to the
   bootstrap stage mainly for Windows-2008/2008r2 since .Net 3.5 is
   required to parse the .json format.

Note - this change has quite a bit of copy/pasta as the packer provider
to write the .json file is repeated in several packer scripts. It is
hoped to compress most of the windows build into a single packer script
in a later ticket/PR which will eliminate much of the duplication.