(IMAGES-1082) Fix WinRM Upload errors
This PR fixes a number of upload issues that were observed when testing
IMAGES-1082 across the full range of builds.

This turned out to be due to sequencing errors introduced with both this
and the previous ticket IMAGES-976 where json parameter files were used
to inject build data.

The Windows "Desktop Feature" which is enabled in the Server editions
appears to be the main culprit as it interacts with the WMF Enable
sequence which allows the packer file uploads to start before the
initial sequence is complete - causing a reboot mid-upload which aborts
the build.

So the original order of the platform-package.ps1 processing has been
reverted to happen during the windows-update processing and the .Net 3.5
enabling which is required for .json processing is brought forward to
the bootstrap phase for Windows 2008/2008r2 only.

Another sequencing weaknesses was also tied down including disabling WMF
before the network is connected (i.e. Hypervisor tools installed).

Also increase the puppet timeout (wait for puppet reboot cycle to
complete) to 60 minutes.