(IMAGES-1095) Win-2008r2/7 SSL Cert Workaround
Adding "AddTrust External CA Root" for Windows 2008/2008r2/7 to overcome
download issues with Artifactory both in the bootstrap phase and during
the Windows 7 puppet run.

The cert is added seamlessly on later versions of Windows - not clear
why this broke now but given the sort remaining lifespan of these OSs
this is an acceptable fix.