(IMAGES-1098) Update common slipstream template
The Slipstream template hasn't followed recent updates in the common
templates, so give it some TLC to bring it up to spec and to get it work
on the vCenter builds.

Also include the boot image .iso name in the json manifest placed on the
image.

The changes are also "copied" to Windows 2008, although this image needs
some additional manual work to generate and finalise the image as DISM
doesn't appear to work satisfactorally in the platform. This platform
is near EOL so this is expected to be the last Slipstream update for it,
but is still considered worth it as it embeds 95 updates in the .iso