(PUP-8085) Set correct Locale for Japanese Tests
Create a new parameter "beaker_locale" to be passed to acceptance test
stage to ensure the beaker tests run with the japananese locale set.
This causes a number of UTF8 related failing tests to be screened out.

Also noted that the French Beakerhostgenerator definitions were
incorrect so have fixed these.