(QENG-7455) Use Puppet code to set firewall
Although this issue turned out to not be firewall related, it still
makes sense to migrate the original powershell firewall commands to
puppet code.

Unfortunately the windows firewall module used had issues on several
platforms so resorting to a simple exec command sequence to test/set
the firewall rules.

Windows 2008/2008r2/7 don't have the Powershell cmdlets, so netsh
for these platforms (which aren't around for much longer anyway)

In both cases, since the command needs to be executed once on the
platform as part of the configuration resort to a simple file marker
to track completion.