(maint) Upgrade Packer to 1.5.1, fix template errors
This commit upgrades the version of packer installed by the `Makefile`
to 1.5.1. After doing that upgrade, I found that several templates now
failed the tests because `_comment` fields were provided as arrays,
which is apparently unsupported now. I fixed up the arrays to be regular
strings and all of the templates appear to pass the tests now.