(IMAGES-1231) November 2020 Windows Image Refresh
Generate the Windows November 2020 Patch Tuesday Image Refresh.

This picks up serious CVE-2020-3941 as well as updates to:
1. Powershell 7.1.0
2. NotePad++ 7.9.1
3. Google chrome 86.0.4240.198
4. Git for Windows 2.29.2.2
5. vmware tools 11.2.0 (via boot process)

Windows 10 is also updated to:
1. win-10-ent: Build 2004
2. win-10-next: Build 2009 (20H2)

Fixed issue with pswindows update testing - seems to be a bug in This
since Build 2004 where the get-wuhistory hangs - adding last 2 to the
command appears to be a vaid workaround.

Google appears to install in the Program Files directory (and no longer
in the x86 one) so adjust in scripting.

Filter our Packer Environment Files on the Forcefully delete so as to
avoid throwing (albeit non-fatal) errors in the cleanup scripts at the
end.

Create PowerCLI script to fixup images for new netapp storage pending
a fix in platform-ci-utils and ci-jobs.

Enter these updates in the command record script.