(PE-31039) Fix refresh_hostname script for 2019.8.4
In 2019.8.4, we set up the PE Infrastructure Agent node group to manage master_uris, pcp_broker_list, and server_list. We need to change the values in here to the new hostname as part of the refresh_hostname script.  Post-2019.8.4, we can add this to the refresh_master_hostname plan so we don't need this patch in the bash file we lay down on the image. But for now, this patch will fix the refresh_hostname.sh script on the 2019.8.4-based image.

Also, somehow, a phantom database node called gummy-chicken managed to sneak into the dumped database, so this purges that node as well.