(REPLATS-168) Add a k8s restart script for kurl image
When a kurl vm is checked out, there's a multi step process to get
containerd and the k8s services running again. It becomes more
complicated when the host ip has changed and k8s services need to be
reconfigured for the new ip.

In order to make it simple for anyone manually checking out a kurl vm to
reset it to a working state, this patch pulls down the restart-k8s.sh
script from the kurl_test repo into /root so it's available to be run
when a vm is checked out.

There doesn't seem to be a straight forward mechanism for including
arbitrary files with puppetlabs-packer without modifying the
templates/common/vmware.vcenter.json to include an additional file
provisioner. But that would affect all the images using that template.