(REPLATS-168) Add provisioners for local git clone and copy
The use case here is to be able to upload files from a separate git
repository into the image.

You can specify a repository to clone onto the builder, and an array of
files to copy from the builder to the image. It's not at all clear from
the template, but you since the file provisioner has the 'sources'
parameter set, you can specify a comma-separated string of files for
custom_files_to_upload, and have it treated as an array per:
https://www.packer.io/docs/templates/legacy_json_templates/user-variables#using-array-values

Your custom script can then move them about on the image and set
permissions as needed.

The provisioner defaults should effectively be noops if not otherwise
set, and therefore not interfere with any existing vcenter vm
builds...hopefully.