(REPLATS-348) Update 8.3-kurl-ha template vars
Fix name, reset version and remove the local-git-checkout step. This was
used for kurl-beta to copy the restart script from the kurl_test repo
onto the image. Since the HA vms will not pre-install kurl, this won't
be necessary.

The kurl.sh script needs to change to a script that just
preloads (or be updated with that option) in a future commit.