(REPLATS-348) Only preload kurl in the 8.3-kurl-ha image
Restarting k8s was challenging for a standalone primary. It's likely to
be worse for an ha primary with ceph/rook, and then we'd need to worry
about joining other primaries to the cluster for ha.

Instead, we're just going to pre-download the kurl tarball, which should
be the bulk of the time, and then install fresh on checked out vms.

Curl's puppet-application-manager-stable instead of standalone, and
updates the kurl.sh to skip installation and managing a restart script
(which we aren't copying into the image) if we are not installing.