(maint) Drain pods before imaging
We're seeing transient issues with pods starting up after reinitializing
via kubeadm. Drain all pods before we stop kubelet and create the image
to ensure pods have a chance to safely shutdown.

Also prevents kubelet and containerd from starting on startup as the
first thing we do in the restart script is to stop them. This avoids
having to cleanup pods again.