(maint) Stop pulling in restart_k8s.sh into kurl image
There's been a lot of churn keeping the image up to date with changes to
the restart script. Manual use of the script (outside of Bolt) is a nice
to have if you are checking out an image for your own use, but not
critical.

Updated motd on the image to provide instructions for obtaining the
script yourself.