(maint) Remove support for installing files from a git checkout
Reverts the general support for checking out an arbitrary git repo
during the build process and copying some subset of files from it into
the image (from 0385f49).

This was only being used in the centos-8.3-kurl image, and since we're
no longer using it there (see previous commit), I'm backing this out to
simplify the templates.