(PT-582) Add a workaround for dnf kurl.local module bug
Kurl's local repositories are configured as dnf modules. This leaves
behind a @modulefailsafe definition for kurl.local tied to the last
local repository kurl installed from. This in turn locks packages that
are in that repository to the repository versions and causes future
upgrade problems.

Calling dnf module reset clears out that stored module metadata,
allowing packages to update normally.