(REPLATS-514) Run k8s-restart after rc-local
The base packer template we use is common/vmware.vcenter.json,
which uses puppet to run the vsphere::nocm class which includes the
packer::vsphere class. That class, on el systems, places the
manifests/modules/packer/templates/vsphere/redhat.rb.erb script on the
image, and sets up /etc/rc.d/rc.local to execute it. This is the script
that is responsible for resetting host name and ip on checkout.

Systemd runs an rc-local.service after network.target that executes
/etc/rc.d/rc.local.

Currently there's a race between k8s-restart and rc-local. k8s-restart
depends on ip being final to work properly, and if it is faster than
rc-local, it will try to use the old ip, or otherwise run into network
errors before dhcp has sorted out dns.

Simply depending on the rc-local.service should fix this. Depending on
both network-online.target and rc-local.service is a bit redundant,
since rc-local.service is depending on network-online.target, but it's
clearer.

Also changed Service Type to oneshot since that appears closer to our
usecase, and added RemainAfterExit true so that systemctl status shows a
green service after exit. It just makes it easier to tell that the
service was successful. This is the same pattern used by systemd's
rc-local.service support.