(REPLATS-616) Workaround bug installing audit-libs in latest kurl-beta
The latest kurl.sh install script has a bug on el8 because of a conflict
between the system audit-libs package and the one being installed by
kurl's containerd.io repo. It can be bypassed by forcing installation of
kurl's package after the first pass of the install fails, and then
reinstalling.

(We need to revert this commit once REPLATS-616 is patched.)