(REPLATS-591) Skip prometheus install on kurl-beta vm
Prometheus monitoring was added back into the standalone pam
configuration. This causes restart to fail because of the hostname reset
and conflicts with the prometheus pvc's node affinity for the previous
'localhost.localdomain' hostname prior to the reset.
(see: puppetlabs/kurl_test@62b39af)

Ultimately, though, we don't really need or want prometheus in the
testing kurl-beta vm, so reintroducing the patch removed in b692f5b to
exclude it from the kurl install.