(maint) Supply default minutes to wait for k8s restart in kurl-beta
Because no default was set for READY_WAIT_MINUTES, the test on line 212
was outputing an error about READY_WAIT_MINUTES not being an integer.
This wasn't preventing the script from running but did add some noise to
the output. This change does also mean that the restart_k8s.service will
fail outright after five minutes if pods aren't running by that time.
That's in line with how kurl_test was handling the restart script back
when the plan was managing this aspect of the vm lifecycle.