(REPLATS-637) Do not wait for terminating pods when restarting kurl-beta
The kurl-beta vm's restart-k8s.sh script has a function to wait for pods
to be in a ready state. This function ignores pods which are already in
the state Completed or Evicted, since they will never be ready, but
there was a timing error in which it might catch a pod in the
Terminating state. And then fail waiting for it when it was no longer
present. Now the script excludes pods in a Terminating state.