(REPLATS-647) Copy audit.yaml back into kurl-beta /etc/kubernetes
kURL v2022.03.25-0 introduces audit configuration for k8s and copies an
audit.yaml file into /etc/kubernetes.

As part of the reset process, we move /etc/kubernetes to
/etc/kubernetes-backup, cp most of the pki back in and let kubeadm
regenerate configuration corrected for the new ip/hostname from
/opt/replicated/kubeadm.conf, which is a sequence of yaml configuration
files for kubeadm.

This process was leaving /etc/kubernetes without an audit.yaml, and this
was causing the kube-apiserver pod to immediately fail.

The audit.yaml does not seem to be recognized by kubeadm, so adding it
to kubeadm.conf does not cause kubeadm to reconstruct it. So the
restart_k8s.sh script now copies the audit.yaml back into
/etc/kubernetes.