(maint) Simplify extra node deletion in kurl-beta restart
A kurl beta checkout ends up with a different hostname and needs to be
reconfigured. This includes deleting the k8s node associated with the
old hostname. This was done conditionally based on whether the restart
script detected a different configured hostname than what `hostname`
reports. But if the restart_k8s.sh has failed after kubeadm as
reconfigured, but before nodes get deleted, subsequent runs of
restart_k8s.sh will fail to delete the old nodes. The heuristic tests
for a node in condition.status Unknown were fragile as well.

To simplify this, I'm just relying on deleting any node that does not
match the current hostname and then waiting for the current hostname
node explicitly.