(maint) Allow kurl-beta restart_k8s.sh to force kotsadm/minio resets
Part of restarting k8s on the kurl-beta vm involves resetting
kotsadm-postgres and minio pods because of their persistent volume's
node affinity for the old deleted node associated with the old hostname.

If the restart fails before this can occur, but after reconfiguration of
k8s for the new hostname, then a subsequent run will skip over this
necessary step.

Since in this state you have to pass a flag to get the script to force
ip/host reset again anyway, I'm also using that flag now to force the
postgres/mino reset.

Typically reusing the restart script doesn't come up much, but it does
matter when debugging failed restarts and re-running.